Split up & move EntityScreen.
