+++ src/components/entity/EntityScreen.js
@@ -3,49 +3,8 @@ import { connect } from 'react-redux';
 
 import Article from 'components/Article';
 import { fetchEntity } from 'actions';
-
-class EntityInfo extends Component {
-  render() {
-    const { entity } = this.props;
-    return (
-      <Article.InfoPane>
-        <h1>{entity.name}</h1>
-        <ul>
-          <li>{entity.schema}</li>
-          <li>{entity.created_at}</li>
-          <li>{entity.countries.join(', ')}</li>
-        </ul>
-      </Article.InfoPane>
-    );
-  }
-}
-
-class EntityContent extends Component {
-  render() {
-    const { entity } = this.props;
-    return (
-      <Article.ContentPane>
-        <dl>
-          {Object.entries(entity.properties).map(([property, values]) => ([
-            <dt>{property}</dt>,
-            <dd>
-              {values.length === 1
-                ? values[0]
-                : (
-                  <ul>
-                    {values.map(value => (
-                      <li>{value}</li>
-                    ))}
-                  </ul>
-                )
-              }
-            </dd>
-          ]))}
-        </dl>
-      </Article.ContentPane>
-    );
-  }
-}
+import EntityInfo from './EntityInfo';
+import EntityContent from './EntityContent';
 
 class EntityScreen extends Component {
   componentDidMount() {
